Update mockEntityList method in MockPerson.kt
The mockEntityList method in MockPerson.kt is modified to accept a quantity parameter that makes it more flexible in generating a list with a specific number of mock Person entities. The hard-coded generation of 13 entities is now adjusted based on passed parameter. This allows better control for various testing scenarios.