Specify return type in PersonServiceTest
The type of returned entity from 'mockEntity' method of 'inputObj' has been explicitly defined as 'Person' in the 'findOne' test case. Previously, the return type was implied and not explicitly stated, which might have led to ambiguity. This change makes easy to understand the expected return type of 'mockEntity' method.