Improve code readability in PersonServiceTest
Removed unused imports, (PersonVO and ArrayList), and unnecessary test 'findAllV2' from PersonServiceTest file. In the 'findOne' test case, additionally checked the size of the result array and confirmed that it is non-null before testing other attributes, enhancing the test case readability and making debugging easier in case of a failure.